feat: Sonarqube properties template and run script
This adds a .sonar directory that contains a template, and an execution
script for running sonarqube scans.

By default, we connect to the red hat sonarqube server. You will need
the project to already exist and a token to access that project.

This is kind of a first pass, but it gets us going.

Signed-off-by: Stephen Adams <tsadams@gmail.com>